ui: rationalise diagram block styling and pre transforms shared by mermaid and svg
